examples/simple-app/pyproject.toml
[project]
dependencies = [
    "typer",
    "python-dotenv",
    "unstructured",
    "langchain-community",
    "langchain-openai",
    "langchain-ollama",
    "langchain-chroma",
    "pyarrow>=17.0.0",
    "fastparquet>=2024.5.0",
]
name = "simple-app"
version = "0.0.1"
description = "A simple app using langchain-graphrag"
authors = [{ name = "Kapil Sachdeva", email = "notan@email.com" }]
readme = "README.md"
requires-python = ">= 3.10"

[tool.rye]
virtual = true
managed = true
dev-dependencies = []
